requirements.txt
langgraph>=0.1.0
langchain>=0.3.0
streamlit>=1.20.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-community>=0.0.30
langchain-google-genai>=2.0.0
langchain-tavily>=0.1.0